refactor(testing): decompose FakePeripheral into FakeConnectionSimulator + FakeGattResponder
